fixes #958 - Clean up repository display per comments.
Modifies the display method in accord with Andy's comments:

NOTE: This is a second pr since we had some real problems with the first one:

1  The ==Repository== line and its corresponding end line surround each
single namespace, and not the entire repo as suggested by their text.
Either change text in these lines, or produce them only once for the
entire repo. DONE. Reformatted the whole output and the functions that
implement it.

2 In the MOF format, the non-MOF lines (e.g. "==...", "Namespace
root/cimv2: contains ...", or "Path=...") could become MOF comment
lines, then the entire output can be used again as MOF.
In the MOF format, the first class in a namespace listing could be
preceded by an empty line.
In the XML format, the non-MOF lines could become XML comment lines
(<!-- ... -->), but I feel that is less important than for MOF (and also
a bit ugly). UGLY or  NOT. DONE

3. In the XML format, the CIM-XML could be pretty-printed.
In the XML format, the objects in a namespace listing could be separated
by empty lines.  DONE

4. In the REPR format, the classes in a namespace listing could be
separated by empty lines. DONE

5. In the XML and MOF formats, the output of the path could use the new
to_wbem_uri() method to make it more explicit that we want a WBEM URI.
AGREED: Really good idea. DONE.  All of the output is surrounded by
comments unless it is the object formatted output to make use of
the ouput simpler.

6. For classes and qualifier decls and namespaces, the output should be
sorted by name, and a TODO should be added that in the future, we should
preserve order.
AGREED. KS. DONE

7. Moves the  testing of the display_repository function to its own
test.