TODO cleanup to eliminate TODOs in FakeWBEMConnection
1. Modified logging. All of logging setup is external to the
mocker and the the mocker simply creates a logger from the
defined mock logger name.  It is the user responsibility to
set up logging for mock.  Note that I also added a try
around the mock methodcalls so we log all exceptions also.

NOTE; This log really duplicates what the WBEMConnection log does but
without the ability to log only portions of the info.

2. Removed the write to file from mofcompiler errors.  Now writes error
information to stdout.

3. Modified the compiler error message format. New format is:

 Syntax error:NoFile:3:24
            Qualifier Association : boolean = false,
                Scope(associations),
                      ^^^^^^^^^^^^
Old format did not put context on separate lines and only showed line
number if there was a file name.  i.e. Showed nothing for String input
to compiler.

Note that the display message still shows error as 'Syntax error' but the
exception is MOFParaseError and that the MOFParseError __str__ output
does not format the context.  The difference is because of the
text in MOFParseError formatter

4.  Moved function to create CIMNamespaceName from Class and instance to
method of cim_obj.CIMInstanceName and created tests. Now static
method in CIMNamespaceName named from_instance(...)

5. Remove some TODOs from FakedWBEMConnection that were not relevent
any more.

6. Cleaned up a couple of pylint issues in test_cim_obj.py but I left
a bunch of others, mostly method attributes that are not used.