Clean test TODOs
Cleaned up the TODOs in the tests and extended a couple of tests to
remove TODOs

Clean up enumerateclasses test and change from_instance

Added lo, iq, ico as parameterized attributes to enumerateclasses test.

Cleaned up tests of get_class to better test ico, iq, lo attributes.

Changed from_instance to ignore any key property not found in instance
if strict=False.  The original behavior was to get the value from the
class but I feel that this is not the correct behavior in general.  this
means that we can produce partial CIMInstanceName keybindings when
strict=False