Test: Renamed pywbem assertion methods to avoid unittest conflicts.
Signed-off-by: Andreas Maier <maiera@de.ibm.com>

Cleaned up run_cimoperations.py PyWBEMListenerClass

1. Built try blocks around all methods that create listeners so we are
certain to always destroy listeners at the end of each test.
2. Build context manager around most uses of Subscription manager to
automatically clean up if we do not clean up manually.
3. Move a number of function to common methods that were being repeated
in individual methos.

The enumerateinstances with deep inheritance test that fails was confirmed
as a server problem in that 1. Instance return includes properties from
superclass that it should not include. 2. Pegasus trace shows correct
value for de and lo getting to server.  Change message slightly to
state that it is just a warning.

Also fixed some issues where the qualifier flavors are now different on
server responses after the latest changes to pywbem.