Remove the logging functionality from pywbem_mock
Completely remove the logging from the mocker since it is little
different from the cim_operations api logger and that logger can be used
to accomplish effectively the same thing. The only real difference
between the two is that that small bits of code that emulates
_imethodcall and _method_call return mapping from pure object to tuples
of data for each response.

Fix flake8 error. Completely remove the try blocks