Bring up-to-date with Andy's comments and clean up code for all of the log configure methods.  This also cleans up the logging docstring in _logging.py
This incorporates Andy's comments that I marked as DONE.  Note that some
of the comments will be reflected in the redesign we agreed on 27 Feb

Change pywbem.ops to pywbem.api in log name per discussion 27 Feb.

Changes for new log config methods in cim_operations

a. Remove the functions from logger except for the configure...from_string
b. Add the methods to WBEMConnection for configure_api_logger,
   configure_http_logger and configue_all_loggers as class methods
c. Remove the enable_log attribute from WBEMConnection
d. Modify the recorder to handle the new detail_level options.
e. Redo many of the tests in both test_logging.py and test_recorder.py

Cleaned up the new methods in cim_operations and created a set of
internal methods since the work is largely common. Added code so that
the class methods can also control the activation of logging.

Tested with wbemcli and cleaned up the output loggers.  In particular
changed the paths logger so that it actually shows the string definition
of paths.

Removed parameter stream in logging.streamhandler()
We had to remove this because apparently not supported in python 2.6