Fixes #1138: Compile mof based on pragma include file from compile_string (#1154)
* Fixes issue where compile_string would not compile a string that
was an include file.  In particular, if the user creates a file that
is just a subset of the schema, this can be written to a disk file and
used to compile those classes, the qualifiers and the dependencies for
those classes (see issue # 1160 for known issue).  However if the same
definition is supplied as a string to compile_string the compile fails
because the compiler simply did not take into account that the input
include file could be a string as well as a file.  There are a couple
minor fixes and one real change when compiling a single pragma include
statement.

Adds 2 tests for MofCompiler to test for these partial schemas and
for what they are expected to compile.

1. test compile of parts of schema from a file.  Used to determine if the
search for missing mof references, etc. works.

2. Tests for compile of MOF string where the mof in the string is mof
set of pragma include statements that represent just a part of the
schema. The test then confirms that these classes and their dependencies
are actually compiled.

* Fix comments about errors, etc.
Fixed issues from Andy's comments and fixed issue in new code for
compile_file where we would output None as file name.  Added tests
specifically for file find errors.