Modify the connection log so that it honors the detail_level=<int> It was always logging the full contexts of the _repr_. Now it outputs up to the integer length if the detail_level is an integer.
This caused the issue Andy noted in issue #1181 to happen in the
tests when I added new test. Moved _response_delay arrtibute in_ init_ in
pywbem_mock/_wbemconnection FakedWBEMConnection to before
superclass init to try to fix this issue. It worked but Andy
please confirm that this was a legitimate fix within Pythonic
rules and stds.

Added more tests to test_recorder class TestLoggingExamples so that
we are testing more of the variations on configure_logger.

fix pylint, etc. issues in test_recorder.py