Add examples from _logging.py as tests. This found a couple more issues in in the recorder code itself which were fixed.
Add new tests and fixe code error in recorder

1. Added several new tests that test WBEMConnection end-end against the
mocker.  These are largely from the  in the _logger.py doc string.

2. Found issue with output of single objects (ex. get_class) with
detail_level = `summary'

3. Found issue where when config_logger has only one logger defined
the other logger is never really setup in recorder.  The detail
level is never reset from None.  Fixed this by adding test to each
logger stage to not log if detail level is None.
This is independent of whether the logger itself is set.

This fixes TestLoggingExamples. (test_5 where basicconfig is set and we
try to log the api for which we did not set the detail level.