Changed default log_dest in config_logger() to None.
In the current code, the default for the 'log_dest' parameter of
'configure_logging()' is "file" because that seemed to be convenient if
one wants to enable logging and uses that log config function for both
Python logger configuration and pywbem activation.

However, this default is inconvenient for users that want to use the
Python logging config functions and use the pywbem log config functions
only for detail level and activation. It seems a bit weird that one needs
to disable the default "file" to None, in that case.

This change therefore changes the default to None. It does that only
for 'configure_logging()' but not for 'configure_logging_from_string()',
because in the string case, one always wants to do both the logger
configuration and the activation of logging for the connection.

Because having a default log filename for 'configure_logging_from_string()'
still makes sense, the default log filename for 'configure_logging()'
remains unchanged.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>