Added a README file for Pypi
This fixes the package information on Pypi that currently shows the
README.rst file but does not format it right:
https://pypi.python.org/pypi/pywbem/0.12.0.

The change introduces a separate README_PYPI.rst file for Pypi,
that has the information we showed there in previous releases,
plus links to the README file on GitHub and to the Documentation
on RTD. These links are versioned to the released version and thus
need to be updated for each release. This step has been added to the
release description in DEVELOP.md.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>