Add capability to pywbem_mock to get DMTF schema, install it, and select classes from the schema to install in a repository.
This adds the class DMTFSchema to pywbem_mock and methods
to get and manage the schema.

For the moment adds a test to get a schema and create a
partial schema.  Note that we probably do not really want
to do this in general because it then requires getting the
schema from the DMTF for each travis, etc. test in the future.
Once this code is confirmed, we will want to only occasionally
use the download part but we can still use the other methods
with the mof in schema

TODO: use this in place of the existing install_dmtf_schema
that is in the testsuite for mof_compiler, etc. testing. Can we
just move that to a separate issue and review this work
on its own???