Fixes #1160: Enhanced MOF compiler and missing dependent classes resolution logic
* Added capability for the mof_compiler when used with a repository
that returns exceptions for the issues  to find reference classnames
and the classes for embedded instances in the compile  search path
if they are not already in the repository.  This works with the
class MOFWBEMConnection.

* This extends MOFWBEMConnection to generate an exception if
the class for a REF defined in a class being compiled does
not exist in the MOFWBEMConnection repository or if the
class defined for an EmbeddedInstance qualifier does not exist.

* The mof_compiler catches these exceptions and tries to compile
in the class by searching the compile search path for the
class and compiling it.

* Note that this capability depends on the behavior of the
target repository. If the repository does not catch the issue or
does not return either CIM_ERR_INVALID_PARAMTER or CIM_ERR_FAILED
the compiler will not search for the missing classes

* Adds tests for cases of these conditions both correct and
error (the REF or EmbeddedInstance class exists or does not exist)
for the conditons in both properties and method parameters.

* Fixes error in one of the test_mofs where the error actually
existed (REF for nonexistent class) and nothing was being caught.

* Adds entry to changes.rst

* Fixed Andy's comments documented in the PR.

* Added diagnostic print for when we add missing dependencies. If we
cannot find them, the result is an exception so this seem logical as
a way for a confused user to figure out why some class that they did
not know about ends up in the repository

* Added simple circular test where the test class is circular to iteslf
throught the EmbeddedInstanceQualifier