Remove old install_dmtf_schema_def code
This extends the first commit by removing the original code from
DMTF_install_schema_def.py and replacing it with the code that
calls the pywbem_mock DMTFSchema class and the definition of the DMTF
schema to use for the pywbem testsuite.  The variables for counts of
qualifiers and classes is kept since that is used only by the compiler
and is linked to the DMTF schema version.

All other tests that used install_dmtf_schema_def were modified to get
variables appropriate for the schema def from the return of the call to
install_dmtf_schema_def rather than predefined variables.  All of the
predefined variables representing parts of the test schema were removed
from the code.