Improved dependencies in makefile.
The makefile is now set up such that partial dependencies are honored and
not rebuilt. This changes the behavior of a number of artifial targets, such
as install or develop, because they are now not always executed, but only
if needed. This allowed adding dependencies t install and/or develop as needed,
so that e.g. make build now can be done without a prior make install.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>