Improved test_function decorator to simplified_test_function
The `pytest_extensions.test_function` decorator was signature-preserving.
This caused the decorated test functions to have unused arguments, and access
to the `kwargs` dict in the test function needed to by key access to the
`kwargs` dict, causing extra code in the test functions.

This change renames the decorator to `simplified_test_function` and improves
the interface of the test functions which now get the testcase tuple as a
named tuple, and the `kwargs` dict of the testcase as expanded keyword
arguments.

Note that the changed decorator needs to explicitly set the `__signature__`
attribute of the test function, because the `signature()` function used by
pytest uses the signature of the unpacked wrapped function.
See also pytest issue #3435.

Increased the minimum versions for development packages: Package "pytest"
from 3.0.7 to 3.3.0 (due to the support for `signature()`) and package
"py" from 1.4.32 to 1.5.1 (because that is the minimum version required by
"pytest").

Signed-off-by: Andreas Maier <maiera@de.ibm.com>