Test: Relaxed warnings check in test_function decorator.
Details:

The current checking in the pytest_extensions.test_function decorator
verified that the warning has been issued exactly once. In some cases
(e.g. deprecated ordering comparisons for NocaseDict), the same
warning message and location is issued more than once (because the
NocaseDict ordering operators delegate to each other). The resulting
warning in the Pythn warning system shows only one warning, because
it merges warnings with the same type, message and location.
However, pytest captures the warnings before that merge and thus
shows more than one warning. This change relaxes the test for
exactly one warning to verify that one or more warnings have been issued
(of the specified type).

An additional difficulty was that when a testcase specifies both an
exception type and a warning type, this is really a bit overspecified,
because the warning could have been issued before the exception
was raised, or not. In theory, one could require that the testcases
have this knowledge, but again it seems a bit overspecified.
This change therefore removes the check for warnings, when an exception
has also been specified in the test case.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>