Fixed Travis issue on py26 with format of empty field error.
Details:

Since 2018-04-27, Travis on Python 2.6 with PACKAGE_LEVEL=latest failed
with 'Error: zero length field name in format' on all testclient .yaml files.

Comparing the Python package versions between 4/26 and 4/27 revealed
only one changed package version: httpretty has increased from 0.8.14
to 0.9.4. It turns out httpretty dropped support for Python 2.6 in
its 0.9 version, and (suspicion) seems to be using format() on
the response we provide in the testclient .yaml files.

This change limits the version of httpretty on Python 2.6 to <0.9.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>