Added PKG-INFO file to support install from GitHub tarball
The source code archives (in .zip and .tar.gz formats) provided on the
GitHub page for a release contain the work directory of the checked out
release. As a result, they are not installable by pip or by setup.py
because the package version cannot be determined by pbr. The installation
fails with this exception:

"Exception: Versioning for this project requires either an sdist tarball,
or access to an upstream git repository. It's also possible that there is
a mismatch between the package name in setup.cfg and the argument given
to pbr.version.VersionInfo. Project name pywbem was given, but was not
able to be found."

The pbr package gets the package version from the first available of the
following sources (see
https://git.openstack.org/cgit/openstack-dev/pbr/tree/pbr/packaging.py):

* Environment variable PBR_VERSION
* Environment variable OSLO_PACKAGE_VERSION
* Git repo (youngest tag in checked out branch, plus devNN since then)
* Metadata file PKG-INFO in main directory (from pywbem.egg-info directory)
* Metadata file METADATA in main directory (from wheel creation)

This change changes the "install" and "build" targets of make to copy
the PKG-INFO file from the directory it is created in (pywbem.egg-info)
to the main directory, and adds the so-created PKG-INFO file to the
git repo. This causes it to be included in the source code archives
built by GitHub which become installable via pip and setup.py as a
result.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>