WIP fixes # 1250
Add a new test for the subscription manager based on pywbem_mock.

NOTE: This upload is WIP and includes temporary code for issue #1251
and the code is prototype level.

This upload is only components of the first test and includes a number
of changes to test_wbemconnection_mock that will be moved to another
repo. They were inserted when trying to sort out issue #1251 under my
assumption that it was a problem in the mock code.