Fix issue with deepcopy of objects in _wbemconnection_mock
We were not using deep copy of objects when we move them to and from the
repository.  Since we are insuring the isolation of the repository we
must do a deep copy on objects retrieved from the repository. We also
do a deep copy on object inserted into the repository with
add_cimobjects just to protect the objects once they are in the
repository.

We made another change. There was a statement that attempted to get
the current default_namespace but actually got the code DEFAULT_NAMESPACE.
This was a bug

We found that the user can cause a problem in trying to access the
instance repo after they have created a class.  The original code
considered this an error in that the instance repo does not exist until
an object is inserted. So it returns the no invalid namespace error and
not the instance not found error  which, while it works, is illogical.
We changed to create the empty instance repo when a class has been added
for that namespace.

Finally, we added one more optional input parameter to
compile_dmtf_schema, 'namespace' since it can be useful to set the
schema into a namespace other than the default namespace.
This is really an extension and not a bug but it is completely backward
compatible.

This required some corrections to the tests, primarily to the definition
of objects for add_cim_object where the lack of deep copy was masking
some differences in parameters returned.

added paragraph to changes.rst

(cherry picked from commit 66e86de43edbef8e645e4db7776841642ec21f3b)

Conflicts:
	pywbem_mock/_wbemconnection_mock.py
	testsuite/test_wbemconnection_mock.py