Fixes #1270, convert wbemcli test to pytest and add more tests
This completely rewrites test_wbemcli.py to use pytest and the same model
for tests as we used for pywbemcli.  Creates a new file
cmd_line_test_utils.py that is the boiler plate for the test and
rewrites test_wbemcli.py to be a single structure of tests and the
method that calls the test code in cmd_line_test_utils.py

This recreates all the tests in the original test_wbemcli.py and also
adds a test for executing a number of the operations as a script.

The tests included are:

1. qualifier decl get and enumerate
2. class enumerate and get
3. instance enumerate and get.

The test for each operation is also part of the script, confirming that
something reasonable is returned.

NOTE: This is not a complete test of all of the options of each of the
operations but does confirm that the method and major parameters do work.

NOTE: This does not test the reference, assocations, invokemethod, or
pull operations.

Fixes issue in wbemcli where help text had extra space at
end of line. This is irrelavent to the working code but the way the
test for help is built, the editor takes that extra space off of
the compare string casusing the tests to fail.