fixes #1271 - Issue with mock EnumerateClassNames and EnumeratClasses
This fixes the issues documented in issue #1271 where there was a param
acquisition name that lower case instead of mixed case (Should have been
ClassName) and a test for invalid classname on EnumerateClassNames and
EnumerateClasses was mising.

This also removes two print statements that should not have been in the
code (printed out the error for the two methods that are not
implemented)
and fixed one pylint issue.

We added tests to cover the new code for ClassName not in repo.

Added note to changes.rst