Fixes #1268 Add input parameter to wbemcli to allow mocking a server (#1269)
* Fixes #1268 Add input parameter to wbemcli to allow mocking a server

This adds a parameter --mock_server to the command line parameters of
wbemcli to allow wbemcli to be used with the pywbem_mock environment.

This also adds one new constructor parameter to FakedWBEMConnection,
timeout that the WBEMConnection can use to return the timeout exception
if this value is set and is of a lower value than the response_delay
time which would indicate that the server delayed longer than the client
timeout setting

It also adds a test of the wbemcli new parameter to test_wbemcli.py and
a simple mof file to be used in that test.

Adds initial tests.  This tests only the mof file compile today

Adds entry under enhancements to changes.rst

Fixed an error in documentation where we use CIMClassName. Should be CIMClass

* Fix Andyy's comments