fixes #1250 Add test_subscription_manager.
This branch defines a new test, test_subscription manager that first creates
the mock repository from the WbemServerMock class and then supplements
this mock repository with the classes required by the subscription manager
and with methods to create subscriptions and test the results as part
of test setup for the subscription tests.

It creates a first prototype test using these changes to test creating
and deleting subscriptions. both owned and not owned

The tests are not complete but provide a framework for more complete tests.

At this point it does only good subscription cases and not errors.