Refactors the test_wbemserverclass and adds two new files: (#1267)
1. instance_from_class_method.py - Temporary localization of the
instance_from_class method until we get it accepted.  This keeps the
tests running until we get this added method accepted as a new method in
ICMInstance.

2. teststuite/wbemserver_mock.py which isolates all the code to generate
the mock for the wbem server.  Since the wbem server mock is required to
run the WbemServer class in other mocks this simplifies building this
component.

3. Modifications to test_wbemserverclass.py to work with the two new
files.

4. In addition we extended the cim_objects we create for the tests
to include attributes that were overlooked before such as propagated and
the description qualifier on some of the objects.