fixes #1188, Add instance_from_class static method, tests and comment in changes.rst
Rebased to pr #1294

Adds new method to CIMInstance, the static method instance_from class
and set of tests for this method.

This also removes the temporary version of this method in testsuite.

NOTE: Currently one test fails and that test is marked FAIL.
We also marked a couple of other tests as FAIL for the moment because
of issue #1188

NOTE: The new method is considerably different internally than the one initially proposed and used in testsuite.

Removes the temp file testsuite/instance_from_class_method.py

Fix issues with test causing failue in python 2.6

Add test for case independence