Improved removal of __pycache__ in make clean
The removal of '__pycache__' directories in the 'clean' target of make
so far piped the list of directories to 'xargs rm -Rf __pycache__'.
The '__pycache__' argument was specified in order to make sure that
'rm -Rf' alone does not behave strangely on some platform (e.g.
by removing everything).
This approach has been changed to pipe the sdifrectory list to
'xargs -n 1 rm -rf' instead, for better clarity.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>