Fixed issue with updating pywbem config variables. (#1308)
Details:

So far, the pywbem config variables were defined in the pywbem.config
namespace and then imported by pywbem into the pywbem namespace.
Pywbem documented that these config variables should be accessed (read
and written) through the pywbem namespace. However, pywbem code
read them in nearly all cases from the pywbem.config namespace.
This caused an update that is done by a pywbem user through the pywbem
namespace, not to be visible in the pywbem.config namespace, so pywbem
did not react to the user's change.

The root cause of this is that in Python, variables are not containers
for values, but names that are bound to objects. Assigning a new value
into the config variable causes that name to be bound to the object
representing the new value.
The variable name in the pywbem namespace is a different name than the
name in the pywbem.config namespace, so binding one of them to a new
object does not change the name binding of the other one.

This change fixes that by only using the pywbem.config namespace
for config variables. They are no longer imported into the
pywbem namespace.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>