Fix issues with case insensitivity on classnames in pywbem_mock. This applies to use of classnames in cmds that must search for subclasses, and that use classnames as parameters (references, etc.)
1. Fix several issues where the case insensitivity in classnames
was not being honored.  This included the methods _get_subclass_names,
_get_class_list_enums, and _classnamelist. Modified to use NocaseDict
where case insensitive tests of classname against iterable was required.

2. Since we did not have alternate case classnames in the tests, we
passed the test suite.  A number of tests were added specifically to
test for case insensitivity for class members on all classname inputs
and the classname on instancename inputs.

3. Extensively cleaned up the tests for reference instance names and
associator instance names since they are what drives the other reference
and associator code

3. We also extended the tests to account for more complex instance
names, in particular in the reference and association tests.

Add more tests and clean up some tests

Refactored use of the string form of CIMError.status_code
Converted this to use the variable rather than the string in all cases.

Add comment to changes.rst