Minor fixes I did not catch in fixes for issue #1355
This removes two commented print statements and also corrects two errors
in the tests.

It also removes a couple of pylint issues and a TODO that had become
invalid.

Finally, it removes the multiple different tests to compare instance names
with expected instance names in favor of the same test in all cases.