This pr adds methods to determine correct issues in the definition of the CIM_ReferencedProfile implemented by DMTF and SNIA SMI.  There is an issue in that they define the use of the antecedent and dependent properties differently.
Adds a new method to WBEMServer determine_cimrefrence_direction(..) that
attempts to determine the definition used by the implementation and
if it does determine the direction, returns a parameter that can be
used with get_central_instances to correct the issue.

For the moment it also implements a parameter in get_central_instances
to bypass the execution of the get_central_instances server method since
that causes issues in some cases.