Added support for pyup.io safety check
The pyup.io website provides a Python package 'safety' which can
be used to run a command line check of the Python package dependencies
against a database of security issues they maintain.

This change adds support for running the pyup.io safety check
to the make 'check' target. A log file 'safety.log' is created
and the log file is also shown on stdout.

At this point, findings in the check do not cause a failure of 'make check'.

The check is run against the package versions specified in the
minimum-constraints.txt file. This required changing the '===' operators
in that file to '=='. The difference between them is subtle and has to do
with treatment of special development versions of packages. Using '=='
should be perfectly fine for our purposes.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>