Extend WBEMConnection_mock to correctly handle resolution of qualifier property, method, parameter from superclass to subclasses. The initial code had a limited resolution method to process the resolution of these entites when a subclass of a class in the repository was created. This change extends the validation of the new class  properties, methods, parameters, and qualifiers so that they are correctly defined superclass properties, methods, etc and the propagated flags are correctly set and moves the resolution to class creation.
Since resolution is required both for compile and add_cimobject, the
change modifies the interaction with the mof compiler to directly set
the output of the compiler into the repository by creating a
new class (MockMOFWBEMConnection based on MOFWBEMConnection) to
resolve the classes when they are inserted into the repository by the
compiler. This subclass to MOFWBEMConnection integrates the
MOFWBEMConnection variables for the class, instance, and qualifier
repositories so that there is no need to copy them back an forth when
the compiler is called.

The code associated with the resolution of classes in separated into
a separate  class (ResolverMixin) that contains just the code to
execute the resolution of CIM classes and their components before
they are inserted into the repository.

Extending the resolution of objects in the repository required changes
to the tests including:

1. Removing some of the parameters on the test objects that are created
(CIM_Foo, etc.) since the resolver that is now part of add_cimobjects
resolves classes  as part of putting these objects into the repository.

2. Extending the tests for CreateClass

The MockMOFWBEMConnection is a subclass of the mof_compiler
MOFWBEMConnection class that:

a. Allows the same definition of the repository to be used by both the
compile_... methods and add_cimobjects methods without merging the
repository instance variables like CIM_Classes, CIM_instances.

b. It overrides the original CreateClass method in MOFWbemConnection to
 resolve each class as it is inserted into the repository by the compiler.

Includes minor modifications to mock documentation.

I also found some errors in the test code (i.e. display of objects, etc)
that were corrected.

NOTE: A test in test_recorder was also modified to enable the logger after
the mock class was compiled since those extra operations
were showing up in the log.

WIP