Specify reference direction in WBEMServer.get_central_instances()
This change adds support for specifying the reference direction
to WBEMServer.get_central_instances() by adding an optional
parameter reference_direction. This was necessary because the DMTF
PRP and the SNIA PRP use the CIM_ReferencedProfile class in different
ways: DMTF PRP defines that the 'Dependent' end goes to the referencing
profile (= autonomous profile), while SNIA PRP defines that the
'Antecedent' end goes to the autonomous profile.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>, Karl Schopmeyer