Adds a test script to the try directory.
Adds a test script currently being used  to test get_central_classes and
the reference difference algorithms to its own branch.

This adds both the script and a json file that defines the characteristics
of many of the registered profiles defined for SNIA SMI-S.

NOTE: This depends on the implementation of get_centralinstances and the difference method defined in pr #1411.

Modify the script to use profile yaml file in end2end in place of the
original json file.

Melded classes ServerDefinitionFile and ServerDefinition from end2end
into this script

1. Melded these classes into this script.  The code uses
ServerDefinition  without change and makes only a single change in
inheriting ServerDefinitionFile (the list_servers method)

2. Extended ServerDefinition class to allow for possible
default_namespace parameter on server definitions.  While this is not
needed for the current tests, since it is a part of the WBEMConnection
object it would seem logical to allow it if there is a chance that this
code may be used more extensively in the future.  Otherwise, we end up
with root/cimv2 as the server default definition and it is clear that
this is not used in general.

Add implementation_namespace to server definition yaml. This was per
discussions and we map that yaml entity to the WBEMConnection default
namespace if it it is provided.  If defined this should be the namespace
that the user would expect the implementation of profiles, etc. to be
in for those cases where the user may not want to go through the scoping
or central methodology to access the profile implementation classes.