Adds a test script to the try directory.
Adds a test script currently being used  to test get_central_classes and
the reference difference algorithms to its own branch.

This adds both the script and a json file that defines the characteristics
of many of the registered profiles defined for SNIA SMI-S.

NOTE: This depends on the implementation of get_centralinstances and the difference method defined in pr #1411.