Fixed location of CIM Schema MOF file archive on DMTF site
The DMTF site makes the MOF files available only under https, and
pywbem was using http. In addition, the alias '/standards/' in the
URL is no longer supported, so that URL segment needed to change to
'/sites/default/files/'.

Because we have checked in the MOF archive, this does not affect
our CI tests, but it would fail if the schema version was upgraded.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>