Extend the manual test run_enum_performance.py to gather more detailed statistical information on the enumerate and corresponsing pull operations and to display more detailed timing information about the time spent in the server vs. the time in the  client.
This not depends on the server time being available to help determine
how much time was spent in the server vs. how much time was spent in
the client parsing the result.

The cmd line interface was expanded to provide for the user to define
test parameters (size of responses, number of responses, size of
pull requests) and also to optionally create a yaml file of the
result.

NOTE: This code is now completely dependent on a WBEM server that
includes a class and method that will generate instance responses with
size and number of instances defined by a method that is part of
that class.  I.e. OpenPegasus.